Some improvements

+++ {{cookiecutter.artifact}}/src/main/java/{{ cookiecutter.package_path }}/biz/auth/permission/request/AddPermissionRequest.java
@@ -48,10 +48,10 @@ public class AddPermissionRequest {
      *
      * @mock 1
      */
-    private Long parentId;
+    private Integer parentId;
 
     /**
-     * 权限类型（1：页面/路由，2：接口/功能 3：按钮/组件）
+     * 权限类型（1：页面/路由，2：接口/功能，3：按钮/组件）
      *
      * @mock 2
      */
